chatbotapp/requirements.txt
openai
langchain
langchain-experimental
python-dotenv
streamlit
huggingface_hub
sqlalchemy